Tests started working
